pyproject.toml
[tool.poetry]
name = "zakaya-leads"
version = "0.1.0"
description = "Lead generation and email automation tool for Zakaya"
authors = ["Matt Toronto <matt@zakaya.io>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
google-auth = "^2.27.0"
google-api-python-client = "^2.118.0"
playwright = "^1.41.0"
langchain = "^0.1.9"
pydantic = "^2.6.1"
requests = "^2.31.0"
asyncio = "^3.4.3"
pandas = "^2.2.0"
sendgrid = "*"
html2text = "*"
google-auth-oauthlib = "^1.2.1"
PyJWT = "^2.6.0"
cryptography = "^42.0.0"
brotli = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 